# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: VTT (Virtual Tabletop) API
  version: 1.2.0
  description: |
    Comprehensive API for Virtual Tabletop platform with real-time collaboration, 
    combat AI, and content management. Supports D&D 5e campaigns with character 
    management, battle maps, AI assistance, and real-time multiplayer sessions.
  contact:
    name: VTT Development Team
    email: api@vtt.dev
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.vtt.example.com
    description: Production server
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from OAuth authentication
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Invalid request parameters
            details:
              type: object
              properties:
                field:
                  type: string
                  example: characterId
                issue:
                  type: string
                  example: Character not found
            timestamp:
              type: string
              format: date-time
            requestId:
              type: string
              example: req_abc123
      required: [error]
...
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
paths:
...
  /auth/logout:
    post:
      tags: [Authentication]
      summary: Logout user
      description: Invalidates the current session
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Successfully logged out
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/me:
    get:
      tags: [Authentication]
      summary: Get current user info
      description: Returns information about the authenticated user
      security:
        - BearerAuth: []
      responses:
        "200":
          description: User information
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  email:
                    type: string
                  displayName:
                    type: string
        "401":
          $ref: "#/components/responses/Unauthorized"

  # Health & Monitoring
...